Moved all logic to a single file that runs the render tests
